Rename package and remove `SecurityConfig` class
Rename to the `authentication` package from `security` package;
Remove `SecurityConfig` to `config` package from `security` package
