+++ src/main/java/com/CezaryZal/authentication/SecurityConstants.java
@@ -1,4 +1,4 @@
-package com.CezaryZal.security;
+package com.CezaryZal.authentication;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
